backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Telegram
telethon==1.34.0
cryptg==0.4.0

# Vector Database
qdrant-client==1.7.0

# ML & NLP
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2
numpy==1.26.3
scikit-learn==1.4.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
orjson==3.9.10
python-dateutil==2.8.2

# Encryption
cryptography==41.0.7

# Socket.io
python-socketio==5.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
